# openapi.yaml
openapi: 3.2.0
info:
  title: SDI - Servicio Distribuido de Inventario
  summary: API de gestion de stock para integracion entre modulos ERP.
  description: '

    API REST del modulo de Stock del sistema SDI.


    Permite administrar productos, consultar stock, crear reservas,

    confirmar ventas, liberar reservas y procesar eventos de compra.


    Esta API esta pensada para integracion entre sistemas internos:

    - Modulo de Ventas

    - Modulo de Compras

    - Modulo de Stock

    - Servicios administrativos internos


    La primera version no incluye seguridad formal. El consumo debe realizarse

    en ambiente controlado de red interna o infraestructura protegida.

    '
  termsOfService: https://docs.sdi.local/terminos-api
  contact:
    name: Equipo SDI - Sistema Distribuido de Inventario
    email: gxbeto@gmail.com
  license:
    name: Uso de la catedra Programacion de Aplicaiciones en Red - Profesor Alfonso
      Fernandez
  version: 0.1.0
servers:
- url: http://127.0.0.1:8000
  description: Ambiente local de desarrollo
- url: http://sdi-pruebas.example.com:6000
  description: Ambiente de pruebas / integracion
paths:
...
  /productos:
    post:
      tags:
      - productos
      summary: Crear producto e inicializar su stock en cero
      description: Registra un producto activo en el catalogo e inserta automaticamente
        su fila en stock_actual con cantidades en cero.
      operationId: crearProducto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoCreate'
      responses:
        '201':
          description: Producto creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoCreated'
              example:
                producto_id: 1
                codigo: INF-006
                estado: CREADO
        '409':
          description: Ya existe un producto con el mismo codigo.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: PRODUCTO_CODIGO_DUPLICADO
                mensaje: Ya existe un producto con el codigo INF-006.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
    get:
      tags:
      - productos
      summary: Listar productos con filtros opcionales
      description: Lista productos del catalogo. Permite filtrar por codigo exacto
        y por estado activo/inactivo.
      operationId: listarProductos
      parameters:
      - name: filtro
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 50
        description: Código exacto del producto. Escriba cualquier código, por ejemplo
          INF-006.
      - name: activo
        in: query
        required: false
        schema:
          type: boolean
        description: Estado del producto. Use true para activos y false para inactivos.
      responses:
        '200':
          description: Lista de productos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoOut'
                title: Response Listar Productos Productos Get
              example:
              - producto_id: 1
                codigo: INF-006
                nombre: Teclado Satellite AK-910 USB / Negro
                descripcion: Teclado USB con cable, color negro
                categoria: Informatica
                unidad_medida: UNIDAD
                impuesto: 10
                activo: true
              - producto_id: 13
                codigo: CEL-001
                nombre: Celular Apple iPhone 17 256GB
                descripcion: Smartphone Apple, 256GB de almacenamiento
                categoria: Celulares
                unidad_medida: UNIDAD
                impuesto: 10
                activo: true
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
  /productos/{producto_id}:
    get:
      tags:
      - productos
      summary: Consultar producto por identificador (consulta flexible agrupada)
      description: 'Devuelve una respuesta agrupada con los bloques solicitados en
        el parámetro `include`: producto, stock, precios y/o movimientos (separados
        por coma). Si no se informa include, se asume `producto`. El orden no afecta
        el resultado y los duplicados se ignoran. Un valor no permitido responde 422
        VALIDACION_ERROR. No reemplaza a GET /stock/{producto_id} ni a GET /stock/movimientos.'
      operationId: obtenerProducto
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
      - name: include
        in: query
        required: false
        schema:
          type: string
          description: 'Bloques a incluir, separados por coma. Valores permitidos:
            producto, stock, precios, movimientos. Ej.: include=producto,stock,precios'
          default: producto
          title: Include
        description: 'Bloques a incluir en la respuesta agrupada, separados por coma.
          Valores permitidos: producto, stock, precios, movimientos. Opcional; si
          se omite se asume ''producto''. El orden no afecta el resultado y los duplicados
          se ignoran. Ej.: include=producto,stock,precios.'
        style: form
        explode: false
      - name: limite_movimientos
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: 'Cantidad máxima de movimientos (solo aplica si include contiene
            ''movimientos''). Orden: más recientes primero.'
          default: 10
          title: Limite Movimientos
        description: 'Cantidad máxima de movimientos (solo aplica si include contiene
          ''movimientos''). Orden: más recientes primero.'
      responses:
        '200':
          description: Producto consultado. Respuesta agrupada según los bloques de
            include.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoConsultaOut'
              examples:
                producto:
                  summary: include=producto
                  value:
                    producto:
                      producto_id: 1
                      codigo: INF-006
                      nombre: Teclado Satellite AK-910 USB / Negro
                      descripcion: Teclado USB con cable, color negro
                      categoria: Informatica
                      unidad_medida: UNIDAD
                      impuesto: 10
                      activo: true
                stock:
                  summary: include=stock
                  value:
                    stock:
                      producto_id: 1
                      cantidad_total: 10
                      cantidad_reservada: 2
                      cantidad_disponible: 8
                productoStockPrecios:
                  summary: include=producto,stock,precios
                  value:
                    producto:
                      producto_id: 1
                      codigo: INF-006
                      nombre: Teclado Satellite AK-910 USB / Negro
                      descripcion: Teclado USB con cable, color negro
                      categoria: Informatica
                      unidad_medida: UNIDAD
                      impuesto: 10
                      activo: true
                    stock:
                      producto_id: 1
                      cantidad_total: 10
                      cantidad_reservada: 2
                      cantidad_disponible: 8
                    precios:
                      precio_compra: 32000
                      precio_venta: 45760
                stockMovimientos:
                  summary: include=stock,movimientos&limite_movimientos=10
                  value:
                    stock:
                      producto_id: 1
                      cantidad_total: 10
                      cantidad_reservada: 2
                      cantidad_disponible: 8
                    movimientos:
                    - movimiento_id: 50
                      producto_id: 1
                      reserva_id: 101
                      tipo_movimiento: VENTA
                      origen: VENTAS
                      documento_ref: FV-001-000123
                      comprobante: STK-001-000001-6
                      cantidad: 2
                      stock_anterior: 10
                      stock_posterior: 8
                      observacion: Confirmación de venta desde reserva
                productoMovimientos:
                  summary: include=producto,movimientos&limite_movimientos=10
                  value:
                    producto:
                      producto_id: 1
                      codigo: INF-006
                      nombre: Teclado Satellite AK-910 USB / Negro
                      descripcion: Teclado USB con cable, color negro
                      categoria: Informatica
                      unidad_medida: UNIDAD
                      impuesto: 10
                      activo: true
                    movimientos:
                    - movimiento_id: 50
                      producto_id: 1
                      reserva_id: 101
                      tipo_movimiento: VENTA
                      origen: VENTAS
                      documento_ref: FV-001-000123
                      comprobante: STK-001-000001-6
                      cantidad: 2
                      stock_anterior: 10
                      stock_posterior: 8
                      observacion: Confirmación de venta desde reserva
        '404':
          description: El producto solicitado no existe.
          content:
            application/json:
              example:
                estado: ERROR
                codigo: PRODUCTO_NO_EXISTE
                mensaje: El producto 1 no existe.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        '403':
          $ref: '#/components/responses/ForbiddenErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        estado:
          type: string
          enum:
          - ERROR
          title: Estado
          description: Estado fijo para respuestas de error.
          default: ERROR
        codigo:
          type: string
          enum:
          - VALIDACION_ERROR
          - PRODUCTO_NO_EXISTE
          - PRODUCTO_INACTIVO
          - PRODUCTO_CODIGO_DUPLICADO
          - STOCK_NO_INICIALIZADO
          - STOCK_INSUFICIENTE
          - RESERVA_NO_EXISTE
          - RESERVA_NO_PENDIENTE
          - RESERVA_NO_COINCIDE
          - NO_AUTENTICADO
          - ACCESO_DENEGADO
          - ERROR_INTERNO
          - ERROR_NO_CONTROLADO
          - MENSAJE_DUPLICADO
          - REPLAY_NO_COINCIDE
          - OPERACION_YA_PROCESADA
          - OPERACION_EN_PROCESO
          - CONFLICTO_CONCURRENCIA
          - DOCUMENTO_YA_PROCESADO
          - DOCUMENTO_YA_TIENE_RESERVA
          - RESERVA_YA_CONFIRMADA
          - RESERVA_YA_LIBERADA
          title: Codigo
          description: Código técnico de error devuelto por la API.
        mensaje:
          type: string
          title: Mensaje
          examples:
          - Descripcion clara del error.
      type: object
      required:
      - codigo
      - mensaje
      title: ErrorResponse
      example:
        codigo: CODIGO_ERROR
        estado: ERROR
        mensaje: Descripcion clara del error.
...
    MovimientoOut:
      properties:
        movimiento_id:
          type: integer
          title: Movimiento Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        reserva_id:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Reserva Id
          description: Debe ser un identificador entero positivo.
        tipo_movimiento:
          type: string
          enum:
          - COMPRA
          - RESERVA
          - VENTA
          - LIBERACION_RESERVA
          - AJUSTE
          title: Tipo Movimiento
          description: Tipo de operación que generó el movimiento de stock.
        origen:
          type: string
          enum:
          - COMPRAS
          - VENTAS
          - STOCK
          - SISTEMA
          title: Origen
          description: Módulo o sistema que originó el movimiento.
        documento_ref:
          anyOf:
          - type: string
            pattern: ^[A-Z]{2,3}-\d{3}-\d{6}$
            description: 'Documento externo recibido desde Ventas o Compras. Se usa
              para trazabilidad externa. Ejemplos: FV-001-000123, OC-001-000555.'
            examples:
            - FV-001-000123
          - type: 'null'
          title: Documento Ref
        comprobante:
          type: string
          pattern: ^STK-\d{3}-\d{6}-\d$
          title: Comprobante
          description: Comprobante interno generado por el módulo Stock para identificar
            la operación dentro del sistema. Incluye dígito verificador para validación
            de entrada.
          examples:
          - STK-001-000001-6
        cantidad:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad
          examples:
          - 10
          example: 10
        stock_anterior:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Stock Anterior
          examples:
          - 10
          example: 10
        stock_posterior:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Stock Posterior
          examples:
          - 10
          example: 10
        observacion:
          anyOf:
          - type: string
          - type: 'null'
          title: Observacion
      type: object
      required:
      - movimiento_id
      - producto_id
      - reserva_id
      - tipo_movimiento
      - origen
      - comprobante
      - cantidad
      - stock_anterior
      - stock_posterior
      - observacion
      title: MovimientoOut
      example:
        cantidad: 2
        comprobante: STK-001-000001-6
        documento_ref: FV-001-000123
        movimiento_id: 50
        observacion: Confirmacion de venta desde reserva
        origen: VENTAS
        producto_id: 1
        reserva_id: 101
        stock_anterior: 10
        stock_posterior: 8
        tipo_movimiento: VENTA
    PrecioResumenOut:
      properties:
        precio_compra:
          anyOf:
          - type: number
            multipleOf: 0.01
            minimum: 0.0
            format: decimal
            example: 32000
          - type: 'null'
          title: Precio Compra
        precio_venta:
          anyOf:
          - type: number
            multipleOf: 0.01
            minimum: 0.0
            format: decimal
            example: 32000
          - type: 'null'
          title: Precio Venta
      type: object
      title: PrecioResumenOut
      example:
        precio_compra: 32000
        precio_venta: 45760
    ProductoConsultaOut:
      properties:
        producto:
          anyOf:
          - $ref: '#/components/schemas/ProductoOut'
          - type: 'null'
        stock:
          anyOf:
          - $ref: '#/components/schemas/StockOut'
          - type: 'null'
        precios:
          anyOf:
          - $ref: '#/components/schemas/PrecioResumenOut'
          - type: 'null'
        movimientos:
          anyOf:
          - items:
              $ref: '#/components/schemas/MovimientoOut'
            type: array
          - type: 'null'
          title: Movimientos
      type: object
      title: ProductoConsultaOut
      description: 'Respuesta agrupada de la consulta flexible GET /productos/{producto_id}.


        Cada bloque se incluye solo si fue solicitado en el parámetro `include`.

        Los bloques no solicitados se omiten de la respuesta.'
      example:
        precios:
          precio_compra: 32000
          precio_venta: 45760
        producto:
          activo: true
          categoria: Informatica
          codigo: INF-006
          descripcion: Teclado USB con cable, color negro
          impuesto: 10
          nombre: Teclado Satellite AK-910 USB / Negro
          producto_id: 1
          unidad_medida: UNIDAD
        stock:
          cantidad_disponible: 8
          cantidad_reservada: 2
          cantidad_total: 10
          producto_id: 1
    ProductoCreate:
      properties:
        codigo:
          type: string
          maxLength: 50
          minLength: 1
          title: Codigo
          examples:
          - INF-006
        nombre:
          type: string
          maxLength: 150
          minLength: 1
          title: Nombre
          examples:
          - Teclado Satellite AK-910 USB / Negro
        descripcion:
          anyOf:
          - type: string
          - type: 'null'
          title: Descripcion
        categoria:
          anyOf:
          - type: string
          - type: 'null'
          title: Categoria
          examples:
          - Informatica
        unidad_medida:
          type: string
          enum:
          - UNIDAD
          - CAJA
          - KG
          - LITRO
          - METRO
          title: Unidad Medida
          description: Unidad de medida utilizada para registrar y mover stock.
          default: UNIDAD
        impuesto:
          type: number
          multipleOf: 0.01
          maximum: 100.0
          minimum: 0.0
          format: decimal
          title: Impuesto
          default: '0.00'
          examples:
          - 10
          example: 10
      type: object
      required:
      - codigo
      - nombre
      title: ProductoCreate
      example:
        categoria: Informatica
        codigo: INF-006
        descripcion: Teclado USB con cable, color negro
        impuesto: 10
        nombre: Teclado Satellite AK-910 USB / Negro
        unidad_medida: UNIDAD
    ProductoCreated:
      properties:
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        codigo:
          type: string
          title: Codigo
        estado:
          type: string
          enum:
          - CREADO
          title: Estado
          description: Estado devuelto cuando el producto fue creado.
          default: CREADO
      type: object
      required:
      - producto_id
      - codigo
      title: ProductoCreated
      example:
        codigo: INF-006
        estado: CREADO
        producto_id: 1
    ProductoOut:
      properties:
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        codigo:
          type: string
          title: Codigo
        nombre:
          type: string
          title: Nombre
        descripcion:
          anyOf:
          - type: string
          - type: 'null'
          title: Descripcion
        categoria:
          anyOf:
          - type: string
          - type: 'null'
          title: Categoria
        unidad_medida:
          type: string
          enum:
          - UNIDAD
          - CAJA
          - KG
          - LITRO
          - METRO
          title: Unidad Medida
          description: Unidad de medida utilizada para registrar y mover stock.
        impuesto:
          type: number
          multipleOf: 0.01
          maximum: 100.0
          minimum: 0.0
          format: decimal
          title: Impuesto
          examples:
          - 10
          example: 10
        activo:
          type: boolean
          title: Activo
      type: object
      required:
      - producto_id
      - codigo
      - nombre
      - descripcion
      - categoria
      - unidad_medida
      - impuesto
      - activo
      title: ProductoOut
      example:
        activo: true
        categoria: Informatica
        codigo: INF-006
        descripcion: Teclado USB con cable, color negro
        impuesto: 10
        nombre: Teclado Satellite AK-910 USB / Negro
        producto_id: 1
        unidad_medida: UNIDAD
...
    StockOut:
      properties:
        producto_id:
          type: integer
          title: Producto Id
          minimum: 1
          description: Debe ser un identificador entero positivo.
        cantidad_total:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad Total
          examples:
          - 10
          example: 10
        cantidad_reservada:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad Reservada
          examples:
          - 10
          example: 10
        cantidad_disponible:
          type: number
          multipleOf: 0.001
          minimum: 0.0
          format: decimal
          title: Cantidad Disponible
          examples:
          - 10
          example: 10
      type: object
      required:
      - producto_id
      - cantidad_total
      - cantidad_reservada
      - cantidad_disponible
      title: StockOut
      example:
        cantidad_disponible: 8
        cantidad_reservada: 2
        cantidad_total: 10
        producto_id: 1
...
  responses:
    ValidationErrorResponse:
      description: El request no cumple el contrato OpenAPI o las validaciones declaradas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: VALIDACION_ERROR
            mensaje: 'campo: descripción clara de la validación incumplida.'
    UnauthorizedErrorResponse:
      description: No autenticado. Falta credencial válida o el mecanismo de autenticación
        rechazó la solicitud.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: NO_AUTENTICADO
            mensaje: No se pudo autenticar la solicitud.
    ForbiddenErrorResponse:
      description: Acceso denegado. El consumidor no tiene permiso para ejecutar la
        operación solicitada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ACCESO_DENEGADO
            mensaje: No tiene permisos para ejecutar esta operación.
    InternalServerErrorResponse:
      description: Error interno controlado del servicio.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ERROR_INTERNO
            mensaje: Ocurrió un error interno al procesar la operación.
    DefaultErrorResponse:
      description: Error no controlado o no clasificado por el contrato.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            estado: ERROR
            codigo: ERROR_NO_CONTROLADO
            mensaje: Ocurrió un error no clasificado por el contrato.
...